# docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This APIs docmantation is only for Swagger Practice",
		"version": "1.0.0",
		"title": "CRUD",
		"contact": "azimpanjwar@gmail.com"
	},
	"host": "localhost:5000",
	"basePath": "/api",
	"schemes": ["http"],
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/user/reset-password/{userId}/{token}": {
			"post": {
				"summary": "user-reset-password",
				"tags": ["userColection"],
				"operationId": "reset-password",
				"deprecated": false,
				"produces": ["application/json"],
				"consumes": ["application/x-www-form-urlencoded"],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"description": "Numeric ID of the user to get",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "token",
						"description": "token which sent on email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "formData",
						"name": "newPassword",
						"description": "Enter newPassword here",
						"type": "string",
						"required": true
					},
					{
						"in": "formData",
						"name": "confirmPassword",
						"description": "Enter confirmPassword here",
						"type": "string",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "changed successfully !! please Login!!",
						"headers": {}
					}
				}
			}
		},
...
	},
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "token",
			"in": "header"
		}
	},
...
}